- Setup keystores for ssl used by ConfigTest.java - Prevented NullPointerException in MemoryPersistenceAdapter
git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/activemq/trunk@386769 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-core/src/main/java/org/apache/activemq/store/memory/MemoryPersistenceAdapter.java
@@ -125,7 +125,10 @@ public class MemoryPersistenceAdapter implements PersistenceAdapter {
                 store.delete();
             }
         }
-        transactionStore.delete();
+
+        if (transactionStore != null) {
+            transactionStore.delete();
+        }
     }
 
     public boolean isUseExternalMessageReferences() {
